ui-components: Move tests to test/ to ensure they are ran
We had some tests file in `lib/ui-components` that were not running at
all. Moving them to `test/` ensures this will not happen again.

Change-type: patch
