+++ Makefile
@@ -248,6 +248,7 @@ lint:
 		lib apps scripts test *.js
 	./scripts/lint/check-filenames.sh
 	./scripts/lint/check-descriptions.sh
+	./scripts/lint/check-tests.sh
 	shellcheck ./scripts/*.sh ./scripts/*/*.sh ./.circleci/*.sh ./deploy-templates/*.sh
 	./node_modules/.bin/deplint
 	./node_modules/.bin/depcheck --ignore-bin-package --ignores='@storybook/*,@babel/*'
